+++ Changelog.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## 1.3.9 - 2020-09-04
+- The previewer now supports up to 64x scaling.
+- After 16x zoom, the pixel grid is automatically displayed.
+
 ## 1.3.8 - 2020-08-01
 - Fixed top bar on svg preview covers top of preview